refactor: CoinsCase 객체 대신 List<CoinCase>를 CoinCaseFactory에서 생성하도록 로직 수정 CoinCase와 인자도 중복되고 단지 CoinCase를 여러개 담는 기능만 있기 때문에 Factory에서 하는 역할이 알맞다고 생각한다.
